+++ Whisky/Views/Setup/GPTKDownloadView.swift
