+++ Whisky/Views/Setup/GPTKInstallView.swift
