+++ Whisky/Views/Setup/RosettaView.swift
