+++ Whisky/Views/Setup/WelcomeView.swift
